tests/requirements.txt
-r ../requirements.txt

coverage
pytest-cov
